requirements.txt
peewee
requests
python-dotenv